requirements.txt
krakenex
pandas
ta
python-dotenv
time
datetime
csv
requests
